Move DataProxy to support dir
