+++ lib/rom/support/data_proxy.rb
