+++ spec/unit/rom/array_dataset_spec.rb
